+++ CSharp_Code/E_Class/E_Class/ProjectFile.cs
@@ -9,7 +9,7 @@ namespace E_Class
     class ProjectFile
     {
         byte[] file;
-        double grade;
+        int grade;
         DateTime uploadDate;
 
 		// Cosntructors
